update mcp server name

+++ _playground/clear-playground.sh
@@ -6,10 +6,11 @@ cd vscode-app-folder
 # Remove all file in playground except .gitignore
 find . -mindepth 1 -not -name '.gitignore' -exec rm -rf {} +
 
-# Save and remove mcp.json
-mkdir -p /tmp/saved-mcp-server-json
-cp ~/.aws/amazonq/mcp.json /tmp/saved-mcp-server-json/mcp.json.$(date +%Y%m%d_%H%M%S)
-rm ~/.aws/amazonq/mcp.json
+# if mcp.json exists
+if [ -f ~/.aws/amazonq/mcp.json ]; then
+    cp ~/.aws/amazonq/mcp.json ~/.aws/amazonq/mcp.json.$(date +%Y%m%d_%H:%M:%S)
+    rm ~/.aws/amazonq/mcp.json
+fi
 
 # Remove Q Developer cache
 if [ "$1" == "--hard" ]; then
